Do what's needed to get Carbon installed in the correct paths
Carbon needs to go in `/opt/graphite/lib` and all its dependencies in
`/opt/graphite/lib/python2.7/site-packages`, or else the packages can't find
each other at import time.

Complicating matters is the fact that, at least for Carbon 0.9.12, the
`requirements.txt` in GitHub contains a different set of dependencies than the
package in PyPi.

To work around these issues, at least for now, we'll install all the
dependencies explicitly and then use a `python_execute` to install Carbon with
all the pip flags it requires.

Newer versions of Carbon look like they have a more accurate `requirements.txt`
so this can potentially be replaced with a `python_requirements` resource when
we're ready to upgrade Graphite and sever the equivalence between the graphite
community cookbook and this one.